requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg[binary]>=3.2
pgvector>=0.3.0
pydantic
python-dotenv
orjson
numpy
pandas
langchain
tiktoken
# Embeddings + reranker
sentence-transformers>=2.7.0
torch
scikit-learn
pytest
httpx